Реализован базовый функционал

+++ tests/conftest.py
@@ -4,34 +4,36 @@ from types import SimpleNamespace
 
 import pytest
 from alembic.command import upgrade
-from messenger.__main__ import parser
 from sqlalchemy import create_engine
-from sqlalchemy.ext.asyncio import create_async_engine, AsyncSession
+from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine
+from sqlalchemy_utils import create_database, database_exists, drop_database
 
-from sqlalchemy.orm import Session
-from sqlalchemy_utils import create_database, drop_database
+from movie_app.__main__ import parser
+from movie_app.api.app import create_app
+from movie_app.constants import Templates
+from movie_app.utils.pg import make_alembic_config
 
-from messenger.api.app import create_app
-from messenger.utils.pg import make_alembic_config
+
+def delete_db(tmp_url):
+    """Удаление БД, если такая существует"""
+    if database_exists(tmp_url):
+        drop_database(tmp_url)
 
 
 @pytest.fixture(scope="session")
 def postgres():
-    """
-    Создает временную БД для запуска теста.
-    """
-    tmp_url = os.environ["TEST_DB"]
+    """Создает временную БД для запуска теста."""
+    tmp_url = os.environ["POSTGRES_DB_URL"]
 
-    try:
-        create_database(tmp_url)
-    except:
-        drop_database(tmp_url)
-        create_database(tmp_url)
+    # Может быть так, что тесты пошли не по плану и БД не удалена,
+    # тогда нужно удалить принудительно
+    delete_db(tmp_url)
+    create_database(tmp_url)
 
     try:
         yield tmp_url
     finally:
-        drop_database(tmp_url)
+        delete_db(tmp_url)
 
 
 @pytest.fixture(scope="session")
@@ -45,19 +47,17 @@ def alembic_config(postgres):
 
 @pytest.fixture(scope="session")
 async def migrated_postgres(alembic_config, postgres):
-    """
-    Возвращает URL к БД с примененными миграциями.
-    """
+    """Возвращает URL к БД с примененными миграциями."""
     upgrade(alembic_config, "head")
     return postgres
 
 
-@pytest.fixture
+@pytest.fixture()
 def arguments(aiomisc_unused_port, migrated_postgres):
     """Аргументы для запуска приложения."""
     return parser.parse_args(
         [
-            "--log-level=debug",
+            "--log-level=info",
             "--api-address=127.0.0.1",
             f"--api-port={aiomisc_unused_port}",
             f"--pg-url={migrated_postgres}",
@@ -67,6 +67,7 @@ def arguments(aiomisc_unused_port, migrated_postgres):
 
 @pytest.fixture()
 async def api_client(aiohttp_client, arguments):
+    """Заводим клиента"""
     app = create_app(arguments)
     client = await aiohttp_client(app, server_kwargs={"port": arguments.api_port})
 
@@ -78,9 +79,7 @@ async def api_client(aiohttp_client, arguments):
 
 @pytest.fixture(scope="session")
 async def migrated_postgres_connection(migrated_postgres):
-    """
-    Синхронное соединение со смигрированной БД.
-    """
+    """Синхронное соединение со смигрированной БД."""
     engine = create_engine(migrated_postgres)
     async with engine.connect() as conn:
         yield conn
@@ -88,11 +87,8 @@ async def migrated_postgres_connection(migrated_postgres):
 
 @pytest.fixture(scope="session")
 async def async_conn(migrated_postgres):
-    """
-    Асинхронное соединение со смигрированной БД.
-    """
-    url = f"postgresql+asyncpg://{os.environ['TEST_DB'].split('://')[1]}"
-    engine = create_async_engine(url)
+    """Асинхронное соединение со смигрированной БД."""
+    engine = create_async_engine(Templates.ASYNC_CONNECTION_URL)
 
     async with AsyncSession(engine) as session:
         async with session.begin():
@@ -102,36 +98,7 @@ async def async_conn(migrated_postgres):
 
 
 @pytest.fixture(scope="session")
-async def async_session(migrated_postgres):
-    """
-    Асинхронное соединение со смигрированной БД.
-    """
-    url = f"postgresql+asyncpg://{os.environ['TEST_DB'].split('://')[1]}"
-    engine = create_async_engine(url)
-
-    yield engine
-
-
-@pytest.fixture(scope="session")
 def loop():
     loop = asyncio.get_event_loop()
     yield loop
     loop.close()
-
-
-@pytest.fixture
-def dbsession(engine, tables):
-    """Returns an sqlalchemy session, and after the test tears down everything properly."""
-    connection = engine.connect()
-    # begin the nested transaction
-    transaction = connection.begin()
-    # use the connection with the already started transaction
-    session = Session(bind=connection)
-
-    yield session
-
-    session.close()
-    # roll back the broader transaction
-    transaction.rollback()
-    # put back the connection to the connection pool
-    connection.close()
